+++ .tryitoutrc.js
@@ -1,0 +1,26 @@
+module.exports = {
+  title: 'cproxy',
+  nav: {
+    Source: 'https://github.com/gabrielcsapo/cproxy'
+  },
+  body: `
+    <div class="text-center">
+      <h3 class="text-black">cproxy</h3>
+      <div class="text-black">⚡️ A high speed proxy cache for all things http</div><br/>
+      <div style="width: 200px;margin:0 auto;">
+        <pre>npm install cproxy -g</pre>
+      </div>
+      <div style="width:60%;margin: 0 auto;">
+        <img class="responsive" src="./example.png" alt="">
+      </div>
+    </div>
+  `,
+  template: 'landing',
+  options: {
+    width: '100%'
+  },
+  output: './docs',
+  footer: `
+    <div class="text-black">Made with ☕️ by <a href="https://www.gabrielcsapo.com">@gabrielcsapo</a></div>
+  `
+}