+++ test/index.js
@@ -4,4 +4,4 @@ test('cproxy', (t) => {
   t.plan(0);
 
   t.end();
-})
+});