+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.1.0 (12/12/2017)
+
+- adds CLI helper
+- adds the option to configure mongo connect string `-m, --mongo [connectString]`
+- adds option to configure the port `-p, --port [port]`
+- adds an extremely simple dashboard
+
 # 0.0.1 (08/30/2017)
 
 - basic use-case